// src/app/api/models/document-type.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Which kind of document this is.
 */
export type DocumentType = 'invoice' | 'receipt' | 'credit_note' | 'quote' | 'estimate' | 'proforma_invoice' | 'purchase_order' | 'statement' | 'timesheet' | 'work_order' | 'packing_slip' | 'sales_order';